fix: form styles
