+++ src/app/screens/Survey/styles.module.scss
@@ -29,11 +29,11 @@
 
 .formContainer {
   background-color: $plain-white;
+  padding: 8px 15px 5px 15px;
 }
 
 .formField {
-  margin: 15px;
-  width: 80%;
+  padding: 15px 0 15px 0;
 }
 
 .surveyButtons {
